Trait type_matrux::client::AuthDrukarnia

pub trait AuthDrukarnia: Deref<Target = Self::Downgrade> {
    type Downgrade;

    // Required methods
    fn authorized_user(&self) -> &AuthorizedUser;
    fn user_set_following<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 UserId,
        follow: bool
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_bookmark_lists<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<FullList>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bookmark_article<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        list: &'life1 ListId,
        article: &'life2 ArticleId
    ) -> Pin<Box<dyn Future<Output = Result<FullBookmark, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn unbookmark_article<'life0, 'life1, 'async_trait>(
        &'life0 self,
        article: &'life1 ArticleId
    ) -> Pin<Box<dyn Future<Output = Result<FullBookmark, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_list_articles<'life0, 'life1, 'async_trait>(
        &'life0 self,
        list: &'life1 ListId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ListArticle>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn like_article<'life0, 'life1, 'async_trait>(
        &'life0 self,
        article: &'life1 ArticleId,
        likes: usize
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn set_comment_liked<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        article: &'life1 ArticleId,
        comment: &'life2 CommentId,
        liked: bool
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

Represents Drukarnia API caller that currently has a valid authenticated user

It’s expected to log user out, once dropped

Required Associated Types

type Downgrade

Represents a type AuthDrukarnia should dereference to

It’s intended to be the type one uses to obtain AuthDrukarnia in the first place

Required Methods

fn authorized_user(&self) -> &AuthorizedUser

Returns authorized user data

fn user_set_following<'life0, 'life1, 'async_trait>( &'life0 self, id: &'life1 UserId, follow: bool ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Implementation details

a request to /api/relationships/subscribe/{USER_ID} POST means “follow” DELETE means “unfollow”

fn get_bookmark_lists<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<FullList>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

GET to /api/articles/bookmarks/lists respond: List of { “_id”:LIST_ID?, “name”:LIST_NAME, “articlesNum”:usize, “owner”:USER_ID, “__v”:0 }

fn bookmark_article<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, list: &'life1 ListId, article: &'life2 ArticleId ) -> Pin<Box<dyn Future<Output = Result<FullBookmark, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

bookmark: POST to /api/articles/bookmarks with json body { “article”: ARTICLE_ID, “list”: LIST_ID } responds with { “article”: ARTICLE_ID, “owner”:OWNER_ID, “list”:LIST_ID, “_id”: BOOKMARK_ID?, “createdAt”: DateTime, “__v”:?? }

un-bookmark: DELETE to /api/articles/{ARTICLE_ID}/bookmarks responds with the same, I guess

fn unbookmark_article<'life0, 'life1, 'async_trait>( &'life0 self, article: &'life1 ArticleId ) -> Pin<Box<dyn Future<Output = Result<FullBookmark, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_list_articles<'life0, 'life1, 'async_trait>( &'life0 self, list: &'life1 ListId ) -> Pin<Box<dyn Future<Output = Result<Vec<ListArticle>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

GET to /api/articles/bookmarks/lists/{LIST_ID} respond: List of { “_id”:ArticleId, “title”:ArticleTitle, “description”:ArticleDescription, “mainTag”:TagName, “readTime”:Duration, “slug”:ArticleSlug, “mainTagSlug”:TagSlug, “mainTagId”:TagId, “createdAt”:DateTime, “isBookmarked”:bool }

fn like_article<'life0, 'life1, 'async_trait>( &'life0 self, article: &'life1 ArticleId, likes: usize ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

like: POST to /api/articles/{ARTICLE_ID}/like with json body { likes: LIKES_NUM???? }

fn set_comment_liked<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, article: &'life1 ArticleId, comment: &'life2 CommentId, liked: bool ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

-like: POST to /api/articles/{ARTICLE_ID}/comments/{COMMENT_ID}/likes with empty json body

unlike: DELETE to same endpoint