Trait type_matrux::client::DrukarniaApi

pub trait DrukarniaApi {
    type Auth;

    // Required methods
    fn popular_tags<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PopularTag>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_user<'life0, 'life1, 'async_trait>(
        &'life0 self,
        name: &'life1 UserName
    ) -> Pin<Box<dyn Future<Output = Result<FullUser, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn search_user_page<'life0, 'life1, 'async_trait>(
        &'life0 self,
        name: &'life1 UserName,
        page: NonZeroUsize
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ShortUser>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_tag<'life0, 'life1, 'async_trait>(
        &'life0 self,
        slug: &'life1 TagSlug
    ) -> Pin<Box<dyn Future<Output = Result<FullTag, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_article<'life0, 'life1, 'async_trait>(
        &'life0 self,
        slug: &'life1 ArticleSlug
    ) -> Pin<Box<dyn Future<Output = Result<FullArticle, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn search_article_page<'life0, 'life1, 'async_trait>(
        &'life0 self,
        name: &'life1 ArticleTitle,
        page: NonZeroUsize
    ) -> Pin<Box<dyn Future<Output = Result<Vec<RecommendedArticle>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_followers_page<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 UserId,
        page: NonZeroUsize
    ) -> Pin<Box<dyn Future<Output = Result<Vec<FollowerUser>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_replies<'life0, 'life1, 'async_trait>(
        &'life0 self,
        comment: &'life1 CommentId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ReplyComment>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn feed_page<'life0, 'async_trait>(
        &'life0 self,
        page: NonZeroUsize
    ) -> Pin<Box<dyn Future<Output = Result<Vec<FeedArticle>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn base_url(&self) -> &Url { ... }
    fn search_user(
        &self,
        name: UserName
    ) -> PageSearchStream<'_, '_, '_, Self::Auth, ShortUser>
       where Self: Sized { ... }
    fn search_article(
        &self,
        name: ArticleTitle
    ) -> PageSearchStream<'_, '_, '_, Self::Auth, RecommendedArticle>
       where Self: Sized { ... }
    fn get_followers(
        &self,
        id: UserId
    ) -> PageSearchStream<'_, '_, '_, Self::Auth, FollowerUser>
       where Self: Sized { ... }
    fn feed(&self) -> PageSearchStream<'_, '_, '_, Self::Auth, FeedArticle>
       where Self: Sized { ... }
    fn login<'life0, 'async_trait>(
        &'life0 self,
        _credentials: Credentials
    ) -> Pin<Box<dyn Future<Output = Result<Self::Auth, Error>> + Send + 'async_trait>>
       where Self::Auth: AuthDrukarnia,
             Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }

}

Represents object’s ability to make requests to Drukarnia’s API.

Required Associated Types

type Auth

Defines a type of this accessor’s authenticated version.

Might even be Self, but this is not desired, as authenticated API accessor is kinda different from a regular one.

Required Methods

fn popular_tags<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<PopularTag>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Retrieves currently popular tags at Drukarnia.

Implementation

Expected to GET /api/articles/tags/popular.

fn get_user<'life0, 'life1, 'async_trait>( &'life0 self, name: &'life1 UserName ) -> Pin<Box<dyn Future<Output = Result<FullUser, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieves a certain user.

Implementation

Expected to GET /api/users/profile/USER_NAME

Errors

Error::NoObject: user with provided name does not exist

fn search_user_page<'life0, 'life1, 'async_trait>( &'life0 self, name: &'life1 UserName, page: NonZeroUsize ) -> Pin<Box<dyn Future<Output = Result<Vec<ShortUser>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Searches a user by it’s name.

Implementation

Expected to GET /api/users/info?name=USER_NAME&page=PAGE&withRelationships=true

fn get_tag<'life0, 'life1, 'async_trait>( &'life0 self, slug: &'life1 TagSlug ) -> Pin<Box<dyn Future<Output = Result<FullTag, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a tag by it’s slug.

Implementation

Expected to GET /api/articles/tags/TAG_SLUG

Errors

Error::NoObject: tag with provided slug does not exists

fn get_article<'life0, 'life1, 'async_trait>( &'life0 self, slug: &'life1 ArticleSlug ) -> Pin<Box<dyn Future<Output = Result<FullArticle, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches an article by it’s slug.

Implementation

Expected to GET /api/articles/ARTICLE_SLUG.

fn search_article_page<'life0, 'life1, 'async_trait>( &'life0 self, name: &'life1 ArticleTitle, page: NonZeroUsize ) -> Pin<Box<dyn Future<Output = Result<Vec<RecommendedArticle>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Searches an article by it’s title.

Implementation

Expected to GET /api/articles/search?name=ARTICLE_TITLE&page=PAGE

fn get_followers_page<'life0, 'life1, 'async_trait>( &'life0 self, id: &'life1 UserId, page: NonZeroUsize ) -> Pin<Box<dyn Future<Output = Result<Vec<FollowerUser>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get followers of a user by it’s id.

Implementation

Expected to GET /api/relationships/USER_ID/followers?page=PAGE

Errors

Error::NoObject: User with provided id does not exist

fn get_replies<'life0, 'life1, 'async_trait>( &'life0 self, comment: &'life1 CommentId ) -> Pin<Box<dyn Future<Output = Result<Vec<ReplyComment>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get replies to a comment.

Implementation

Expected to GET /api/articles/ARTICLE_ID/comments/COMMENT_ID/replies, but at the time of me writing this, result is independent of ARTICLE_ID part, and thus can be retrieved with it being just zeroes:

GET /api/articles/000000000000000000000000/comments/COMMENT_ID/replies

fn feed_page<'life0, 'async_trait>( &'life0 self, page: NonZeroUsize ) -> Pin<Box<dyn Future<Output = Result<Vec<FeedArticle>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a single feed page.

Implementation

Expected to GET to /api/preferences/feed?page=PAGE

Provided Methods

fn base_url(&self) -> &Url

Returns base url accessor uses to call DrukarniaAPI.

https://drukarnia.com.ua/ by default.

fn search_user( &self, name: UserName ) -> PageSearchStream<'_, '_, '_, Self::Auth, ShortUser>where Self: Sized,

Searches user by it’s name.

Returns

A stream of search result pages.

Implementation details

This function should not be reimplemented.

Currently, the underlying stream consequently calls for search result pages, although this might be changed in the future for more optimized approach.

Note

Stream ends after first error, since after the error, there’s no way for stream to determine, if search results had ended

fn search_article( &self, name: ArticleTitle ) -> PageSearchStream<'_, '_, '_, Self::Auth, RecommendedArticle>where Self: Sized,

Searches an article by it’s title.

Returns

A stream of search result pages.

Implementation details

This function should not be reimplemented.

Currently, the underlying stream consequently calls for search result pages, although this might be changed in the future for more optimized approach.

Note

Stream ends after first error, since after the error, there’s no way for stream to determine, if search results had ended

fn get_followers( &self, id: UserId ) -> PageSearchStream<'_, '_, '_, Self::Auth, FollowerUser>where Self: Sized,

Get followers of a user by it’s id.

Returns

A stream of result pages.

Implementation details

This function should not be reimplemented.

Currently, the underlying stream consequently calls for result pages, although this might be changed in the future for more optimized approach.

Note

Stream ends after first error, since after the error, there’s no way for stream to determine, if results had ended

fn feed(&self) -> PageSearchStream<'_, '_, '_, Self::Auth, FeedArticle>where Self: Sized,

Get articles at feed.

Returns

A stream of result pages.

Implementation details

This function should not be reimplemented.

Currently, the underlying stream consequently calls for result pages, although this might be changed in the future for more optimized approach.

Note

Stream ends after first error, since after the error, there’s no way for stream to determine, if results had ended

fn login<'life0, 'async_trait>( &'life0 self, _credentials: Credentials ) -> Pin<Box<dyn Future<Output = Result<Self::Auth, Error>> + Send + 'async_trait>>where Self::Auth: AuthDrukarnia, Self: Sync + 'async_trait, 'life0: 'async_trait,

Logs in a Drukarnia user.

Currently, should not be used and/or implemented