Module type_matrux::object

Contains data objects as well as some infrastructure for their definition and usage.

All of the provided objects are purposely immutable - this way you can guarantee that actual data stored inside them was indeed returned from a site. For validation purposes, they all have [get_age] function, returning [time::Duration] representing time passed since their creation.

Structs

ArticleComment

ArticleDescription

Represents a description of some article

ArticleId

Represents an id of some article

ArticleSlug

Represents a slug of some article

ArticleTag

ArticleTitle

Represents a title of some article

ArticleUser

AuthorArticle

AuthorizedUser

BookmarkId

Represents an id of some bookmark

CommentId

Represents an id of some article comment

CommentUser

Credentials

Represents user credentials

FeedArticle

FollowerUser

FullArticle

FullBookmark

FullList

FullTag

FullUser

ListArticle

ListId

Represents an id of some user list

PopularTag

RecommendedArticle

Relationships

Represents user’s attitude to some object (other user, tag, article, etc)

ReplyComment

SearchArticle

ShortArticle

ShortUser

TagArticle

TagId

Represents an id of some tag

TagName

Represents a name of some tag

TagSlug

Represents a slug of some tag

UserDescription

Represents a description of some user

UserDisplayName

Represents a display name of some user

UserId

Represents an id of some user

UserName

Represents a username of some user

UserShortDescription

Represents a short description of some user

UserSocials

User’s social links, like telegram and facebook

UserTag

Enums

MaybeUrl

Url type for all of the requests